move addresses to utils
