--- src/addrs/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod addresses;
-pub(crate) use addresses::{Mac, Bssid};
