+++ src/builders/packets/ether_header.rs
@@ -1,4 +1,4 @@
-use crate::addrs::Mac;
+use crate::utils::Mac;
