+++ src/builders/packets/tcp_builder.rs
@@ -1,6 +1,6 @@
 use std::net::Ipv4Addr;
-use crate::addrs::Mac;
 use crate::builders::packets::{ip_header, ether_header, Checksum};
+use crate::utils::Mac;
