+++ src/engines/flood_beacon/beacon_flood.rs
@@ -2,12 +2,11 @@ use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 use std::time::Instant;
 use crate::engines::BcFloodArgs;
-use crate::addrs::Bssid;
 use crate::iface::{Iface, IfaceManager};
 use crate::builders::Frames;
 use crate::sockets::Layer2Socket;
 use crate::generators::RandomValues;
-use crate::utils::{ CtrlCHandler, abort};
+use crate::utils::{CtrlCHandler, abort, Bssid};
