+++ src/engines/wifi_map/wifi_data.rs
@@ -1,5 +1,5 @@
 use std::collections::HashSet;
-use crate::addrs::Bssid;
+use crate::utils::Bssid;
