+++ src/main.rs
@@ -1,4 +1,3 @@
-mod addrs;
 mod builders;
 mod dissectors;
 mod generators;
